+++ verl/workers/megatron_workers.py
@@ -418,6 +418,31 @@ class ActorRolloutRefWorker(MegatronWorker):
         assert self._is_actor
         pass
 
+    @register(dispatch_mode=Dispatch.ONE_TO_ALL, execute_mode=Execute.RANK_ZERO)
+    def push_to_hub(self, repo_id: str, local_path: str, private: bool = False, token: str = None):
+        """Push a model to the Hugging Face Hub.
+
+        Args:
+            repo_id (`str`):
+                The name of the repository you want to push your model to. It should be in the format `username/model_name`.
+            local_path (`str`):
+                Path to the local directory containing the model files.
+            private (`bool`, *optional*, defaults to `False`):
+                Whether the model repository should be private.
+            token (`str`, *optional*):
+                The token to use as HTTP bearer authorization for remote files. If None, will use the token generated when
+                running `huggingface-cli login` (stored in `~/.huggingface`).
+        """
+        from huggingface_hub import HfApi
+
+        api = HfApi()
+        api.create_repo(repo_id=repo_id, private=private, token=token, exist_ok=True)
+        api.upload_folder(
+            folder_path=local_path,
+            repo_id=repo_id,
+            token=token,
+        )
+
 
 class CriticWorker(MegatronWorker):
